Edited March 18
In .step() of customEnv of class,
First, I focused on "done" which is satisfied only when the number of epsidoe equals 200.

In addition, reward function, which is one of the most important notion in RL, is complete. The concept called 'global variable' in action_memory, step_count made me code easier.

Finally, Dimension of reward function was also changed from matrix to scalar, which is reasonable because the reward function is a function of an input step t.